Fix Polymarket API decoding for optional metadata
- Align DTO fields and types with official API contracts
- Handle string thresholds and correct resolutionSource mapping
- Share JSON settings between clients and regression tests
- Cover missing metadata, market selection, and display ordering
- Document the API contract audit and remaining compatibility risks

+++ app/src/main/java/com/streamatico/polymarketviewer/data/model/gamma_api/BaseMarketDto.kt
@@ -24,7 +24,7 @@ interface BaseMarketDto {
     val closed: Boolean
     val isArchived: Boolean
 
-    val groupItemThreshold: Int?
+    val groupItemThreshold: Double?
     val groupItemTitle: String?
 
     val closedTime: OffsetDateTime?
